requirements.txt
Flask==1.1.1
SQLAlchemy==1.3.6
python-dotenv==0.10.3
python-telegram-bot==11.1.0